+++ operator-framework/src/test/java/io/javaoperatorsdk/operator/baseapi/statuscache/internal/StatusPatchCacheStatus.java
@@ -1,0 +1,15 @@
+package io.javaoperatorsdk.operator.baseapi.statuscache.internal;
+
+public class StatusPatchCacheStatus {
+
+  private Integer value = 0;
+
+  public Integer getValue() {
+    return value;
+  }
+
+  public StatusPatchCacheStatus setValue(Integer value) {
+    this.value = value;
+    return this;
+  }
+}
